test: isolate sweep status coverage
